fix(maps): extract getTopicMarkerColor to client-safe map-utils.ts
map-queries.ts는 Prisma(→ pg → dns) 의존성으로 클라이언트 번들 불가.
getTopicMarkerColor/resolveTopicColorHex를 Prisma 없는 map-utils.ts로 분리.
ExploreMapView, MapPageClient 모두 @/lib/map-utils 에서 import.
